Add files via upload
Code that  gets  either  H or  T  from the  user.  This  can be  a chain  of multiple  h's and  t's. e.g HTHTHTHTHHHHT
If there is only  only one H or T inputted and  if its H then  it automatically ends the code and  says success
Otherwise it will  go through the algorithm.
step 1.  find the last H in the input.
step 2.  find the letter behind  and  in front  of  it and filp it. H = T || T = H.
Step 3. Remove the  last H in the input.
Step 4. Repeat until one  letter is left