Add files via upload
A palindrome is a word that can be spelled backwards but remains the same. For example, "racecar" spelled backwards is still "racecar," and "radar" remains "radar." The code checks for any numbers or symbols, then converts everything to lowercase. It also removes spaces before reversing the word and comparing it to the original.